+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- Changed `tree-sitter-hl-add-patterns` to support language-specific patterns, in addition to buffer-local patterns.
 
 ## [0.8.3] - 2020-07-12
 - Fixed incorrect highlighting when region-to-highlight's boundaries cut query patterns in halves.